+++ src/viriatum/test/simple_test.c
@@ -25,6 +25,7 @@
 #include "stdafx.h"
 
 #include "simple_test.h"
+#include "handler_dispatch_test.h"
 #include "handler_file_test.h"
 
 #ifndef VIRIATUM_NO_THREADS
@@ -1026,6 +1027,7 @@ void exec_simple_tests(struct test_case_t *test_case) {
     V_RUN_TEST(test_handler_file_url, test_case);
     V_RUN_TEST(test_handler_file_header_field, test_case);
     V_RUN_TEST(test_handler_file_header_value, test_case);
+    V_RUN_TEST(test_dispatch_handler_context_keepalive, test_case);
 }
 
 ERROR_CODE run_simple_tests() {